fetch roles eficintly
