+++ gcba_users/app/globals.css
@@ -32,7 +32,7 @@ input[type="number"]::-webkit-inner-spin-button,
     margin: 0;
   }
 
-input, textarea {
+input, textarea, select {
   background-color: white;
   color: #0a0a0a;
   border-radius: 8px;
@@ -40,7 +40,7 @@ input, textarea {
   resize: none;
 }
 
-input, textarea:focus {
+input, select, textarea:focus {
   border: 0px;
   outline: #ffffff;
 }
