+++ gcba_users/prisma/schema.prisma
@@ -21,6 +21,7 @@ model Employee {
 model Role {
   id        Int        @id @default(autoincrement())
   value     String     @unique
+  name      String
   employees Employee[]
   createdAt DateTime   @default(now())
   updatedAt DateTime   @updatedAt
